CHROMIUM: Add eclass variables to portage lists
This is a port of the eclass variables from the existing
shellcheck-0.6.0-ebuild-vars.patch (see crrev.com/c/1373917 for details)
to 0.7. The list of known variables is the same, but they go into the
isPortageBuild lists instead of the global list.  A possible side effect
of this change is that non-ebuild files using any of these names as
globals will now show SC2034 errors, but this is a good outcome if it
happens.

BUG=chromium:989299
TEST=shellcheck shellcheck-0.6.0.ebuild

Change-Id: I8cfdc6e2abe285816f87747c48f0050b451062e2
Reviewed-on: https://chromium-review.googlesource.com/1783468
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Ready: Benjamin Gordon <bmgordon@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Will Bradley <wbbradley@chromium.org>