CHROMIUM: stop complaining about 'which' in portage build scripts
This fix is fairly blunt. It simply disables checkWhich if we are in the
context of a Portage Build file. A slightly better direction might be to
allow each individual check to determine it's activation based on
parameters, however the interface for checkers is Token -> Analysis. And
thus, it does not have the context of the params at that point. We can
plumb the params down later once there are more instances of this sort
of check pruning.

BUG=chromium:980463
TEST=stack build && stack exec -- shellcheck --source-path=SCRIPTDIR <some .ebuild with usage of which> && stack test

Change-Id: I179b26e3f2ba55980925932bff9430d9c4aa38e2
Reviewed-on: https://chromium-review.googlesource.com/1817198
Tested-by: Will Bradley <wbbradley@chromium.org>
Commit-Ready: Will Bradley <wbbradley@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>