CHROMIUM: Extract arguments from T_Redirecting
The current arguments function only handles T_SimpleCommand, but
top-level commands are actually wrapped in a T_Redirecting.  This causes
crashes with "non-exhaustive pattern match" when parsing ebuilds.

Add a case that extracts the arguments from the inner command.  Since
`getCommand` handles all types of tokens, we can always attempt to
extract the inner command for everything that isn't a T_SimpleCommand.

BUG=chromium:1002679
TEST=shellcheck /mnt/host/src/.../chromeos-config-bsp-grunt-9999.ebuild

Change-Id: Ifef0e8d7578767ad22fd84a6424598891fc9ee05