CHROMIUM: Add semantics for tc-export functions
tc-export creates and exports a list of build variables.  For example,
`tc-export CC` should be treated as an assignment to CC as well as a
reference to it.  Similarly, tc-export_build_env creates an explicit
list plus several additional copies of *FLAGS build variables.

shellcheck already has a concept of commands that represent a variable
reference as well as a variable assignment, so let's add the tc-export
variants to both of these.

BUG=b:187789914
TEST=stack test; cros lint ghostscript-gpl-9.56.1.ebuild

Change-Id: If0aad1561da6a7653286f3cbca693d873584075c